+++ src/master/detail.js
